requirements.txt
Flask
flask-bcrypt
flask-cors
flask-jwt-extended
psycopg2-binary
